RwsTest.Convertors/XmlToJsonConverter.cs
using Newtonsoft.Json;
using RwsTest.Common.Exceptions;
using RwsTest.Common.Interfaces;
using System;
using System.Xml;

namespace RwsTest.Convertors
{
    public class XmlToJsonConverter : IFormatConverter
    {
        /// <summary>
        /// Converts xml to json
        /// </summary>
        /// <param name="input"></param>
        /// <returns></returns>
        public string Convert(string input)
        {
            if (string.IsNullOrWhiteSpace(input?.Trim()))
            {
                throw new ArgumentNullException(nameof(input));
            }

            var doc = new XmlDocument();

            try
            {
                doc.LoadXml(input);

                return JsonConvert.SerializeXmlNode(doc, Newtonsoft.Json.Formatting.Indented);
            }
            catch (Exception ex)
            {
                throw new ConversionFailedException("Conversion failed", ex);
            }
        }
    }
}
